# src/hotel-management-api-spec.yaml
openapi: 3.0.3
info:
  title: Hotel Management System API
  description: |
    RESTful API for Hotel Operations Digitalization - Room Booking & Reservation Module.
    This API enables hotel staff to manage room availability, reservations, guest check-in/check-out, billing, and payments.
  version: 1.0.0
  contact:
    name: Hotel Management System Team
    email: support@hotelmanagementsystem.com
servers:
  - url: https://api.hotelmanagementsystem.com/v1
    description: Production server
  - url: https://staging-api.hotelmanagementsystem.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with credentials and return JWT token
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: frontdesk_user
                password:
                  type: string
                  format: password
                  example: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT authentication token
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
                  expiresIn:
                    type: integer
                    description: Token expiration time in seconds
                    example: 1800
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current user session
      operationId: logoutUser
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /check-in:
    post:
      tags:
        - Check-In/Check-Out
      summary: Check-in guest
      description: Process guest check-in for a confirmed reservation
      operationId: checkInGuest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - confirmationNumber
              properties:
                confirmationNumber:
                  type: string
                  example: CNF-2024-001234
                idProofVerified:
                  type: boolean
                  example: true
                specialRequests:
                  type: string
                  example: Extra pillows requested
      responses:
        '200':
          description: Check-in successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservation:
                    $ref: '#/components/schemas/Reservation'
                  roomKey:
                    type: string
                    example: KEY-101-2024-001
                  checkInTime:
                    type: string
                    format: date-time
                    example: '2024-03-15T14:30:00Z'
        '400':
          description: Check-in validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                earlyCheckIn:
                  value:
                    code: EARLY_CHECK_IN
                    message: Cannot check in before reservation start date
                    timestamp: '2024-01-15T10:30:00Z'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /check-out:
    post:
      tags:
        - Check-In/Check-Out
      summary: Check-out guest
      description: Process guest check-out and finalize billing
      operationId: checkOutGuest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - confirmationNumber
              properties:
                confirmationNumber:
                  type: string
                  example: CNF-2024-001234
                checkOutTime:
                  type: string
                  format: date-time
                  example: '2024-03-18T11:45:00Z'
      responses:
        '200':
          description: Check-out successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservation:
                    $ref: '#/components/schemas/Reservation'
                  finalBill:
                    $ref: '#/components/schemas/Bill'
                  checkOutTime:
                    type: string
                    format: date-time
                    example: '2024-03-18T11:45:00Z'
                  lateCheckOutCharge:
                    type: number
                    format: double
                    example: 0.00
        '400':
          description: Check-out validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unpaidBill:
                  value:
                    code: UNPAID_BILL
                    message: Payment must be settled before check-out completion
                    timestamp: '2024-01-15T10:30:00Z'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /billing/{confirmationNumber}:
    get:
      tags:
        - Billing
      summary: Get bill details
      description: Retrieve itemized bill for a reservation
      operationId: getBillByConfirmationNumber
      parameters:
        - name: confirmationNumber
          in: path
          required: true
          schema:
            type: string
            example: CNF-2024-001234
      responses:
        '200':
          description: Bill retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /billing/{confirmationNumber}/calculate:
    post:
      tags:
        - Billing
      summary: Calculate bill
      description: Calculate total bill including room charges, services, and taxes
      operationId: calculateBill
      parameters:
        - name: confirmationNumber
          in: path
          required: true
          schema:
            type: string
            example: CNF-2024-001234
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                additionalCharges:
                  type: array
                  items:
                    $ref: '#/components/schemas/Charge'
      responses:
        '200':
          description: Bill calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /payments:
    post:
      tags:
        - Billing
      summary: Record payment
      description: Record guest payment for a reservation
      operationId: recordPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /payments/{confirmationNumber}:
    get:
      tags:
        - Billing
      summary: Get payment history
      description: Retrieve payment history for a reservation
      operationId: getPaymentHistory
      parameters:
        - name: confirmationNumber
          in: path
          required: true
          schema:
            type: string
            example: CNF-2024-001234
      responses:
        '200':
          description: Payment history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
                  totalPaid:
                    type: number
                    format: double
                    example: 15000.00
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /cancellations:
    post:
      tags:
        - Cancellations
      summary: Cancel reservation
      description: Cancel a reservation and apply cancellation policy charges
      operationId: cancelReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - confirmationNumber
              properties:
                confirmationNumber:
                  type: string
                  example: CNF-2024-001234
                reason:
                  type: string
                  example: Guest requested cancellation due to change in plans
      responses:
        '200':
          description: Reservation cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  confirmationNumber:
                    type: string
                    example: CNF-2024-001234
                  cancellationId:
                    type: string
                    example: CXL-2024-001234
                  cancellationCharge:
                    type: number
                    format: double
                    example: 1000.00
                  refundAmount:
                    type: number
                    format: double
                    example: 4000.00
                  cancellationPolicy:
                    type: string
                    example: 20% cancellation charge for cancellations within 48 hours of check-in
                  cancellationTime:
                    type: string
                    format: date-time
                    example: '2024-03-13T09:15:00Z'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /cancellations/{cancellationId}:
    get:
      tags:
        - Cancellations
      summary: Get cancellation details
      description: Retrieve details of a cancelled reservation
      operationId: getCancellationDetails
      parameters:
        - name: cancellationId
          in: path
          required: true
          schema:
            type: string
            example: CXL-2024-001234
      responses:
        '200':
          description: Cancellation details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancellationId:
                    type: string
                    example: CXL-2024-001234
                  confirmationNumber:
                    type: string
                    example: CNF-2024-001234
                  cancellationCharge:
                    type: number
                    format: double
                    example: 1000.00
                  refundAmount:
                    type: number
                    format: double
                    example: 4000.00
                  reason:
                    type: string
                    example: Guest requested cancellation due to change in plans
                  cancellationTime:
                    type: string
                    format: date-time
                    example: '2024-03-13T09:15:00Z'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
    User:
      type: object
      properties:
        userId:
          type: string
          example: USR-001
        username:
          type: string
          example: frontdesk_user
        role:
          type: string
          enum: [ADMIN, FRONT_DESK_STAFF, MANAGER]
          example: FRONT_DESK_STAFF
        fullName:
          type: string
          example: John Smith
        email:
          type: string
          format: email
          example: john.smith@hotel.com
...
    Guest:
      type: object
      required:
        - name
        - contactNumber
        - email
        - idProofType
        - idProofNumber
      properties:
        guestId:
          type: string
          example: GST-001234
        name:
          type: string
          example: John Doe
        contactNumber:
          type: string
          example: '+91-9876543210'
        email:
          type: string
          format: email
          example: john.doe@email.com
        idProofType:
          type: string
          enum: [PASSPORT, DRIVERS_LICENSE, NATIONAL_ID, AADHAAR]
          example: PASSPORT
        idProofNumber:
          type: string
          example: P1234567
        address:
          type: string
          example: 123 Main Street, City, Country
...
    Reservation:
      type: object
      properties:
        confirmationNumber:
          type: string
          example: CNF-2024-001234
        roomId:
          type: string
          example: RM-101
        roomNumber:
          type: string
          example: '101'
        roomType:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DELUXE
        checkInDate:
          type: string
          format: date
          example: '2024-03-15'
        checkOutDate:
          type: string
          format: date
          example: '2024-03-18'
        guest:
          $ref: '#/components/schemas/Guest'
        numberOfGuests:
          type: integer
          example: 2
        numberOfNights:
          type: integer
          example: 3
        status:
          type: string
          enum: [CONFIRMED, CHECKED_IN, CHECKED_OUT, CANCELLED]
          example: CONFIRMED
        specialRequests:
          type: string
          example: Late check-in expected
        createdAt:
          type: string
          format: date-time
          example: '2024-03-10T14:30:00Z'
        checkInTime:
          type: string
          format: date-time
          nullable: true
          example: '2024-03-15T14:30:00Z'
        checkOutTime:
          type: string
          format: date-time
          nullable: true
          example: '2024-03-18T11:45:00Z'
    Charge:
      type: object
      properties:
        chargeId:
          type: string
          example: CHG-001
        description:
          type: string
          example: Room Service
        amount:
          type: number
          format: double
          example: 500.00
        chargeType:
          type: string
          enum: [ROOM_CHARGE, SERVICE_CHARGE, TAX, LATE_CHECKOUT, OTHER]
          example: SERVICE_CHARGE
        chargeDate:
          type: string
          format: date-time
          example: '2024-03-16T18:30:00Z'
    Bill:
      type: object
      properties:
        billId:
          type: string
          example: BILL-2024-001234
        confirmationNumber:
          type: string
          example: CNF-2024-001234
        charges:
          type: array
          items:
            $ref: '#/components/schemas/Charge'
        roomCharges:
          type: number
          format: double
          example: 15000.00
        serviceCharges:
          type: number
          format: double
          example: 2000.00
        taxAmount:
          type: number
          format: double
          example: 2040.00
        lateCheckOutCharge:
          type: number
          format: double
          example: 0.00
        totalAmount:
          type: number
          format: double
          example: 19040.00
        paymentStatus:
          type: string
          enum: [PENDING, PARTIAL, PAID]
          example: PENDING
        generatedAt:
          type: string
          format: date-time
          example: '2024-03-18T11:45:00Z'
    PaymentRequest:
      type: object
      required:
        - confirmationNumber
        - amount
        - paymentMethod
      properties:
        confirmationNumber:
          type: string
          example: CNF-2024-001234
        amount:
          type: number
          format: double
          example: 19040.00
        paymentMethod:
          type: string
          enum: [CASH, CREDIT_CARD, DEBIT_CARD, UPI]
          example: CREDIT_CARD
        transactionReference:
          type: string
          example: TXN-CC-123456789
    Payment:
      type: object
      properties:
        paymentId:
          type: string
          example: PAY-2024-001234
        confirmationNumber:
          type: string
          example: CNF-2024-001234
        amount:
          type: number
          format: double
          example: 19040.00
        paymentMethod:
          type: string
          enum: [CASH, CREDIT_CARD, DEBIT_CARD, UPI]
          example: CREDIT_CARD
        transactionReference:
          type: string
          example: TXN-CC-123456789
        paymentStatus:
          type: string
          enum: [SUCCESS, FAILED, PENDING]
          example: SUCCESS
        paymentTime:
          type: string
          format: date-time
          example: '2024-03-18T11:50:00Z'
        receiptNumber:
          type: string
          example: RCP-2024-001234
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request parameters
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        details:
          type: array
          items:
            type: string
          example: ["checkInDate must be in the future", "checkOutDate must be after checkInDate"]
  responses:
    BadRequestError:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: BAD_REQUEST
            message: Invalid request parameters
            timestamp: '2024-01-15T10:30:00Z'
            details: ["checkInDate is required"]
    UnauthorizedError:
      description: Unauthorized - Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired authentication token
            timestamp: '2024-01-15T10:30:00Z'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Reservation not found
            timestamp: '2024-01-15T10:30:00Z'
...